rewrote the clipboard in typescript
We fixed several issues with hidding the elemtn as well as ensuring child content inside of a clipboard target didn't interfere with the click event
