+++ package.json
@@ -1,8 +1,8 @@
 {
   "name": "aurelia-clipboard",
-  "version": "1.0.2",
+  "version": "1.0.3",
   "description": "aurelia custom attribute for using browser's clipboard api",
-  "main": "src/clipboard.js",
+  "main": "src/clipboard.ts",
   "scripts": {
     "test": "echo \"Error: no test specified\" && exit 1"
   },
@@ -21,5 +21,8 @@
   "bugs": {
     "url": "https://github.com/devanp92/aurelia-clipboard/issues"
   },
-  "homepage": "https://github.com/devanp92/aurelia-clipboard#readme"
+  "homepage": "https://github.com/devanp92/aurelia-clipboard#readme",
+  "dependencies": {
+    "aurelia-framework": "^1.1.4"
+  }
 }
